dont crash on defaults
